+++ pyuploadcare/api/__init__.py
@@ -1,7 +1,9 @@
 from pyuploadcare.api.api import (  # noqa: F401
+    AddonsAPI,
     DocumentConvertAPI,
     FilesAPI,
     GroupsAPI,
+    MetadataAPI,
     ProjectAPI,
     UploadAPI,
     VideoConvertAPI,