+++ pyuploadcare/api/entities.py
@@ -1,10 +1,13 @@
+import re
 from datetime import datetime
 from decimal import Decimal
 from enum import Enum
-from typing import Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple, TypeVar, Union
 from uuid import UUID
 
-from pydantic import BaseModel, EmailStr, PrivateAttr
+from pydantic import BaseModel, ConstrainedStr, EmailStr, Field, PrivateAttr
+
+from .metadata import META_KEY_MAX_LEN, META_KEY_PATTERN, META_VALUE_MAX_LEN
 
 
 class Entity(BaseModel):
@@ -80,15 +83,108 @@ class VideoInfo(Entity):
     duration: Decimal
     format: str
     bitrate: Decimal
-    audio: Optional[AudioStreamInfo]
-    video: VideoStreamInfo
+    audio: List[AudioStreamInfo]
+    video: List[VideoStreamInfo]
+
+
+class MIMEInfo(Entity):
+    mime: str
+    type: str
+    subtype: str
+
+
+class ContentInfo(Entity):
+    mime: Optional[MIMEInfo]
+    image: Optional[ImageInfo]
+    video: Optional[VideoInfo]
+
+
+class ApllicationDataDetails(Entity):
+    pass
+
+
+DetailsType = TypeVar(
+    "DetailsType", bound=ApllicationDataDetails, covariant=True
+)
+
+
+class ApplicationDataBase(Entity):
+    data: Optional[Union[Dict[str, Any], ApllicationDataDetails]] = Field(
+        default_factory=dict
+    )
+    version: str
+    datetime_created: datetime
+    datetime_updated: datetime
+
+
+class AWSRecognitionLabel(Entity):
+    confidence: Decimal = Field(alias="Confidence")
+    instances: List[Dict] = Field(alias="Instances")
+    name: str = Field(alias="Name")
+    parents: List[Dict] = Field(alias="Parents")
+
+
+class AWSRecognitionDetectLabelsDetails(ApllicationDataDetails):
+    label_model_version: Optional[str] = Field(alias="LabelModelVersion")
+    labels: List[AWSRecognitionLabel] = Field(
+        alias="Labels", default_factory=list
+    )
+
+
+class AWSRecognitionDetectLabelsApplicationData(ApplicationDataBase):
+    data: AWSRecognitionDetectLabelsDetails
+
+
+class RemoveBackgroundDetails(ApllicationDataDetails):
+    foreground_type: Optional[str]
+
+
+class RemoveBackgroundApplicationData(ApplicationDataBase):
+    data: RemoveBackgroundDetails
+
+
+class UCClamAVDetails(ApllicationDataDetails):
+    infected: Optional[bool]
+    infected_with: Optional[str]
+
+
+class UCClamAVApplicationData(ApplicationDataBase):
+    data: UCClamAVDetails
+
+
+class ApplicationDataSet(Entity):
+    aws_rekognition_detect_labels: Optional[
+        AWSRecognitionDetectLabelsApplicationData
+    ]
+    remove_bg: Optional[RemoveBackgroundApplicationData]
+    uc_clamav_virus_scan: Optional[UCClamAVApplicationData]
+
+
+class MetadataKeyConStrType(ConstrainedStr):
+    regex = re.compile(META_KEY_PATTERN)
+    max_length = META_KEY_MAX_LEN
+
+
+class MetadataValueConStrType(ConstrainedStr):
+    max_length = META_VALUE_MAX_LEN
+
+
+MetadataDict = Dict[MetadataKeyConStrType, MetadataValueConStrType]
 
 
 class FileInfo(UUIDEntity):
+    """
+    video_info, image_info and rekognition_info were deprecated in REST API v0.7
+
+    video_info, image_info moved to content_info
+    rekognition_info moved to app_data
+
+    """
+
     datetime_removed: Optional[datetime]
     datetime_stored: Optional[datetime]
     datetime_uploaded: Optional[datetime]
-    image_info: Optional[ImageInfo]
+    metadata: Optional[MetadataDict]
     is_image: Optional[bool]
     is_ready: Optional[bool]
     mime_type: Optional[str]
@@ -97,9 +193,9 @@ class FileInfo(UUIDEntity):
     size: Optional[int]
     url: Optional[str]
     variations: Optional[Dict[str, UUID]]
-    video_info: Optional[VideoInfo]
     source: Optional[str]
-    rekognition_info: Optional[Dict[str, Decimal]]
+    content_info: Optional[ContentInfo]
+    appdata: Optional[ApplicationDataSet]
 
 
 class GroupInfo(Entity):